This was unslothai's quantize.cpp circa Q2 2025.
They haven't updated it since upstream moved the examples into tools
folder so I put it here, not sure if anything else was in the commits.

It doesn't compiles given the upstream changes.

Personally, I prefer the `--custom-q` and `--tensor-type` methods
peronsonally as it doesn't depend on brittle hard codes.